requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.6
# apturl==0.5.2
asgiref==3.4.0
async-timeout==4.0.1
attrs==21.2.0
authclient==1.0
authentication==1.1.0
autopep8==1.5.7
billiard==3.6.4.0
blinker==1.4
# Brlapi==0.7.0
celery==5.1.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
# command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
# cupshelpers==1.0
dbus-python==1.2.16
# defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.4.0
# distro-info===0.23ubuntu1
Django==3.2.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==21.1
django-import-export==2.6.1
django-js-asset==1.2.2
django-redis==5.0.0
django-rest-passwordreset==1.2.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
# djangorestframework-simplejwt==4.7.2
dnspython==2.1.0
drf-yasg==1.20.0
ecdsa==0.17.0
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.1.0
Flask==2.0.1
frozenlist==1.2.0
future==0.18.2
# gpg===1.13.1-unknown
greenlet==1.1.2
httpie==2.4.0
httplib2==0.14.0
idna==2.8
inflection==0.5.1
intuit-oauth==1.2.4
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
keyring==18.0.1
kombu==5.1.0
# language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
# louis==3.12.0
macaroonbakery==1.3.1
MarkupPy==1.14
MarkupSafe==2.0.1
multidict==5.2.0
netifaces==0.10.4
# numpy==1.21.2
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
# pandas==1.3.3
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
prompt-toolkit==3.0.20
protobuf==3.6.1
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.10.4
pycups==1.9.73
Pygments==2.9.0
# PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
# python-apt==2.0.0+ubuntu0.20.4.6
python-crontab==2.5.1
python-dateutil==2.7.3
# python-debian===0.1.36ubuntu1
python-jose==3.3.0
python-quickbooks==0.9.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rauth==0.7.3
redis==3.5.3
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.17.6
six==1.14.0
sqlparse==0.4.1
# systemd-python==234
tablib==3.1.0
toml==0.10.2
typing_extensions==4.0.0
# ubuntu-advantage-tools==27.2
# ubuntu-drivers-common==0.0.0
# ufw==0.36
# unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
# xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
